rename CR alternative to CR
