+++ src/components/pages/collectRecordFormPages/CollectRecordFormPage/getObservationValidationInfo.js
