+++ src/components/pages/collectRecordFormPages/CollectRecordFormPage/index.js
@@ -1,0 +1,3 @@
+import CollectRecordFormPage from './CollectRecordFormPage'
+
+export default CollectRecordFormPage
