--- src/components/pages/collectRecordFormPages/CollectRecordFormPageAlternative/index.js
@@ -1,3 +1,0 @@
-import CollectRecordFormPageAlternative from './CollectRecordFormPageAlternative'
-
-export default CollectRecordFormPageAlternative
